+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release Versions:
 - Remove type from SpatialState constructor (#3)
 - Reorganize function declarations and implementations (#4, #8)
 - Remove empty flag from State and SpatialState constructors (#7)
+- Refactor State::is_compatible to is_incompatible (#6)
 
 ## 6.3.1